+++ example/todo/run-in-node.js
@@ -1,0 +1,2 @@
+React = require("react");
+require("../../dist/build/todo-node/todo-node.jsexe/all.js");